Casi terminada la migracion

+++ src/main/java/com/proyecto/dao/CorreoDao.java
@@ -9,9 +9,7 @@ import com.proyecto.entities.Bootcamper;
 import com.proyecto.entities.Correo;
 
 @Repository
-public interface CorreoDao extends JpaRepository <Correo,Integer> {
-
-
+public interface CorreoDao extends JpaRepository<Correo, Long> {
     long deleteByBootcamper (Bootcamper bootcamper);    
     List<Correo> findByBootcamper (Bootcamper bootcamper);
